modified AndroidManifest.xml to incorporate an Application class, and added said class
