+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,7 @@
         android:maxSdkVersion="32" />
 
     <application
+        android:name="StudyBuddies"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
