+++ app/src/main/java/com/github/se/studybuddies/MainActivity.kt
@@ -39,6 +39,7 @@ import com.google.firebase.auth.FirebaseAuth
 class MainActivity : ComponentActivity() {
   private lateinit var auth: FirebaseAuth
 
+
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     auth = FirebaseAuth.getInstance()
