+++ app/src/main/java/com/github/se/studybuddies/StudyBuddies.kt
@@ -1,0 +1,10 @@
+package com.github.se.studybuddies
+
+import android.app.Application
+
+class StudyBuddies : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+    }
+}
