+++ app/src/main/java/com/github/se/studybuddies/navigation/NavigationActions.kt
@@ -18,6 +18,9 @@ class NavigationActions(private val navController: NavHostController) {
   }
 
   companion object {
-    fun navigateTo(route: String) {}
+    fun navigateTo(route: String) {
+
+    }
   }
+
 }
