+++ app/src/main/java/com/github/se/studybuddies/viewModels/ToDoViewModel.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
 class ToDoViewModel(private val uid: String? = null) : ViewModel() {
-  private val firebaseConnection = DatabaseConnection()
+
   private val _todo = MutableStateFlow(emptyToDo())
   val todo: StateFlow<ToDo> = _todo
